separate aws.go!
