--- aws.go
@@ -1,566 +1,0 @@
-package stefunny
-
-import (
-	"context"
-	"errors"
-	"fmt"
-	"log"
-	"strings"
-
-	"github.com/aws/aws-sdk-go-v2/aws"
-	"github.com/aws/aws-sdk-go-v2/service/cloudwatchlogs"
-	"github.com/aws/aws-sdk-go-v2/service/eventbridge"
-	eventbridgetypes "github.com/aws/aws-sdk-go-v2/service/eventbridge/types"
-	"github.com/aws/aws-sdk-go-v2/service/sfn"
-)
-
-type CloudWatchLogsClient interface {
-	cloudwatchlogs.DescribeLogGroupsAPIClient
-}
-type EventBridgeClient interface {
-	PutRule(ctx context.Context, params *eventbridge.PutRuleInput, optFns ...func(*eventbridge.Options)) (*eventbridge.PutRuleOutput, error)
-	ListRuleNamesByTarget(ctx context.Context, params *eventbridge.ListRuleNamesByTargetInput, optFns ...func(*eventbridge.Options)) (*eventbridge.ListRuleNamesByTargetOutput, error)
-	DescribeRule(ctx context.Context, params *eventbridge.DescribeRuleInput, optFns ...func(*eventbridge.Options)) (*eventbridge.DescribeRuleOutput, error)
-	ListTargetsByRule(ctx context.Context, params *eventbridge.ListTargetsByRuleInput, optFns ...func(*eventbridge.Options)) (*eventbridge.ListTargetsByRuleOutput, error)
-	PutTargets(ctx context.Context, params *eventbridge.PutTargetsInput, optFns ...func(*eventbridge.Options)) (*eventbridge.PutTargetsOutput, error)
-	DeleteRule(ctx context.Context, params *eventbridge.DeleteRuleInput, optFns ...func(*eventbridge.Options)) (*eventbridge.DeleteRuleOutput, error)
-	RemoveTargets(ctx context.Context, params *eventbridge.RemoveTargetsInput, optFns ...func(*eventbridge.Options)) (*eventbridge.RemoveTargetsOutput, error)
-	ListTagsForResource(ctx context.Context, params *eventbridge.ListTagsForResourceInput, optFns ...func(*eventbridge.Options)) (*eventbridge.ListTagsForResourceOutput, error)
-	TagResource(ctx context.Context, params *eventbridge.TagResourceInput, optFns ...func(*eventbridge.Options)) (*eventbridge.TagResourceOutput, error)
-}
-
-var (
-	ErrScheduleRuleDoesNotExist = errors.New("schedule rule does not exist")
-	ErrRuleIsNotSchedule        = errors.New("this rule is not schedule")
-)
-
-type ScheduleRule struct {
-	eventbridge.PutRuleInput
-	TargetRoleArn string
-	Targets       []eventbridgetypes.Target
-}
-
-type ScheduleRules []*ScheduleRule
-
-type EventBridgeService interface {
-	DescribeScheduleRule(ctx context.Context, ruleName string, optFns ...func(*eventbridge.Options)) (*ScheduleRule, error)
-	SearchScheduleRule(ctx context.Context, stateMachineArn string) (ScheduleRules, error)
-	DeployScheduleRules(ctx context.Context, rules ScheduleRules, optFns ...func(*eventbridge.Options)) (DeployScheduleRulesOutput, error)
-	DeleteScheduleRules(ctx context.Context, rules ScheduleRules, optFns ...func(*eventbridge.Options)) error
-}
-
-var _ EventBridgeService = (*EventBridgeServiceImpl)(nil)
-
-type EventBridgeServiceImpl struct {
-	client EventBridgeClient
-}
-
-func NewEventBridgeService(client EventBridgeClient) *EventBridgeServiceImpl {
-	return &EventBridgeServiceImpl{
-		client: client,
-	}
-}
-
-func (svc *EventBridgeServiceImpl) DescribeScheduleRule(ctx context.Context, ruleName string, optFns ...func(*eventbridge.Options)) (*ScheduleRule, error) {
-	describeOutput, err := svc.client.DescribeRule(ctx, &eventbridge.DescribeRuleInput{Name: &ruleName}, optFns...)
-	if err != nil {
-		if strings.Contains(err.Error(), "ResourceNotFoundException") {
-			return nil, ErrScheduleRuleDoesNotExist
-		}
-		return nil, err
-	}
-	log.Println("[debug] describe rule:", MarshalJSONString(describeOutput))
-	if describeOutput.ScheduleExpression == nil {
-		return nil, ErrRuleIsNotSchedule
-	}
-	listTargetsOutput, err := svc.client.ListTargetsByRule(ctx, &eventbridge.ListTargetsByRuleInput{
-		Rule:  &ruleName,
-		Limit: aws.Int32(5),
-	}, optFns...)
-	if err != nil {
-		return nil, err
-	}
-	log.Println("[debug] list targets by rule:", MarshalJSONString(listTargetsOutput))
-	tagsOutput, err := svc.client.ListTagsForResource(ctx, &eventbridge.ListTagsForResourceInput{
-		ResourceARN: describeOutput.Arn,
-	}, optFns...)
-	if err != nil {
-		return nil, err
-	}
-	rule := &ScheduleRule{
-		PutRuleInput: eventbridge.PutRuleInput{
-			Name:               describeOutput.Name,
-			Description:        describeOutput.Description,
-			EventBusName:       describeOutput.EventBusName,
-			EventPattern:       describeOutput.EventPattern,
-			RoleArn:            describeOutput.RoleArn,
-			ScheduleExpression: describeOutput.ScheduleExpression,
-			State:              describeOutput.State,
-			Tags:               tagsOutput.Tags,
-		},
-		Targets: listTargetsOutput.Targets,
-	}
-	return rule, nil
-}
-
-type ListStateMachineAliasesPaginator struct {
-	client    SFnClient
-	params    *sfn.ListStateMachineAliasesInput
-	nextToken *string
-	firstPage bool
-}
-
-func newListStateMachineAliasesPaginator(client SFnClient, params *sfn.ListStateMachineAliasesInput) *ListStateMachineAliasesPaginator {
-	if params == nil {
-		params = &sfn.ListStateMachineAliasesInput{}
-	}
-
-	return &ListStateMachineAliasesPaginator{
-		client:    client,
-		params:    params,
-		firstPage: true,
-	}
-}
-
-func (p *ListStateMachineAliasesPaginator) HasMorePages() bool {
-	return p.firstPage || p.nextToken != nil
-}
-
-func (p *ListStateMachineAliasesPaginator) NextPage(ctx context.Context, optFns ...func(*sfn.Options)) (*sfn.ListStateMachineAliasesOutput, error) {
-	if !p.HasMorePages() {
-		return nil, fmt.Errorf("no more pages available")
-	}
-
-	params := *p.params
-	params.NextToken = p.nextToken
-
-	result, err := p.client.ListStateMachineAliases(ctx, &params, optFns...)
-	if err != nil {
-		return nil, err
-	}
-	p.firstPage = false
-
-	prevToken := p.nextToken
-	p.nextToken = result.NextToken
-
-	if prevToken != nil && p.nextToken != nil && *prevToken == *p.nextToken {
-		p.nextToken = nil
-	}
-	return result, nil
-}
-
-type ListStateMachineVersionsPaginator struct {
-	client    SFnClient
-	params    *sfn.ListStateMachineVersionsInput
-	nextToken *string
-	firstPage bool
-}
-
-func newListStateMachineVersionsPaginator(client SFnClient, params *sfn.ListStateMachineVersionsInput) *ListStateMachineVersionsPaginator {
-	if params == nil {
-		params = &sfn.ListStateMachineVersionsInput{}
-	}
-
-	return &ListStateMachineVersionsPaginator{
-		client:    client,
-		params:    params,
-		firstPage: true,
-	}
-}
-
-func (p *ListStateMachineVersionsPaginator) HasMorePages() bool {
-	return p.firstPage || p.nextToken != nil
-}
-
-func (p *ListStateMachineVersionsPaginator) NextPage(ctx context.Context, optFns ...func(*sfn.Options)) (*sfn.ListStateMachineVersionsOutput, error) {
-	if !p.HasMorePages() {
-		return nil, fmt.Errorf("no more pages available")
-	}
-
-	params := *p.params
-	params.NextToken = p.nextToken
-
-	result, err := p.client.ListStateMachineVersions(ctx, &params, optFns...)
-	if err != nil {
-		return nil, err
-	}
-	p.firstPage = false
-
-	prevToken := p.nextToken
-	p.nextToken = result.NextToken
-
-	if prevToken != nil && p.nextToken != nil && *prevToken == *p.nextToken {
-		p.nextToken = nil
-	}
-	return result, nil
-}
-
-type listRuleNamesByTargetPaginator struct {
-	client    EventBridgeClient
-	params    *eventbridge.ListRuleNamesByTargetInput
-	nextToken *string
-	firstPage bool
-}
-
-func newListRuleNamesByTargetPaginator(client EventBridgeClient, params *eventbridge.ListRuleNamesByTargetInput) *listRuleNamesByTargetPaginator {
-	if params == nil {
-		params = &eventbridge.ListRuleNamesByTargetInput{}
-	}
-
-	return &listRuleNamesByTargetPaginator{
-		client:    client,
-		params:    params,
-		firstPage: true,
-	}
-}
-
-func (p *listRuleNamesByTargetPaginator) HasMorePages() bool {
-	return p.firstPage || p.nextToken != nil
-}
-
-func (p *listRuleNamesByTargetPaginator) NextPage(ctx context.Context, optFns ...func(*eventbridge.Options)) (*eventbridge.ListRuleNamesByTargetOutput, error) {
-	if !p.HasMorePages() {
-		return nil, fmt.Errorf("no more pages available")
-	}
-
-	params := *p.params
-	params.NextToken = p.nextToken
-
-	result, err := p.client.ListRuleNamesByTarget(ctx, &params, optFns...)
-	if err != nil {
-		return nil, err
-	}
-	p.firstPage = false
-
-	prevToken := p.nextToken
-	p.nextToken = result.NextToken
-
-	if prevToken != nil && p.nextToken != nil && *prevToken == *p.nextToken {
-		p.nextToken = nil
-	}
-	return result, nil
-}
-
-func (svc *EventBridgeServiceImpl) SearchScheduleRule(ctx context.Context, stateMachineArn string) (ScheduleRules, error) {
-	log.Printf("[debug] call SearchScheduleRule(ctx,%s)", stateMachineArn)
-	p := newListRuleNamesByTargetPaginator(svc.client, &eventbridge.ListRuleNamesByTargetInput{
-		TargetArn: aws.String(stateMachineArn),
-	})
-	rules := make([]*ScheduleRule, 0)
-	for p.HasMorePages() {
-		output, err := p.NextPage(ctx)
-		if err != nil {
-			return nil, err
-		}
-		for _, name := range output.RuleNames {
-			log.Println("[debug] detect rule: ", name)
-			schedule, err := svc.DescribeScheduleRule(ctx, name)
-			if err != nil && err != ErrRuleIsNotSchedule {
-				return nil, err
-			}
-			if err == ErrRuleIsNotSchedule {
-				continue
-			}
-			if schedule.IsManagedBy() {
-				rules = append(rules, schedule)
-			} else {
-				name := ""
-				if schedule.Name != nil {
-					name = *schedule.Name
-				}
-				log.Printf("[debug] found a scheduled rule `%s` that %s does not manage.", name, appName)
-			}
-		}
-	}
-	log.Printf("[debug] end SearchScheduleRule() %d rules found", len(rules))
-	return rules, nil
-}
-
-type DeployScheduleRuleOutput struct {
-	RuleArn          *string
-	FailedEntries    []eventbridgetypes.PutTargetsResultEntry
-	FailedEntryCount int32
-}
-
-func (svc *EventBridgeServiceImpl) DeployScheduleRule(ctx context.Context, rule *ScheduleRule, optFns ...func(*eventbridge.Options)) (*DeployScheduleRuleOutput, error) {
-	log.Println("[debug] deploy put rule")
-	putRuleOutput, err := svc.client.PutRule(ctx, &rule.PutRuleInput, optFns...)
-	if err != nil {
-		return nil, err
-	}
-	log.Println("[debug] deploy put targets")
-	putTargetsOutput, err := svc.client.PutTargets(ctx, &eventbridge.PutTargetsInput{
-		Rule:    rule.Name,
-		Targets: rule.Targets,
-	}, optFns...)
-	if err != nil {
-		return nil, err
-	}
-
-	log.Println("[debug] deploy update tag")
-	rule.AppendTags(map[string]string{
-		tagManagedBy: appName,
-	})
-	_, err = svc.client.TagResource(ctx, &eventbridge.TagResourceInput{
-		ResourceARN: putRuleOutput.RuleArn,
-		Tags:        rule.PutRuleInput.Tags,
-	})
-	if err != nil {
-		return nil, err
-	}
-	output := &DeployScheduleRuleOutput{
-		RuleArn:          putRuleOutput.RuleArn,
-		FailedEntries:    putTargetsOutput.FailedEntries,
-		FailedEntryCount: putTargetsOutput.FailedEntryCount,
-	}
-	return output, nil
-}
-
-type DeployScheduleRulesOutput []*DeployScheduleRuleOutput
-
-func (o DeployScheduleRulesOutput) FailedEntryCount() int32 {
-	total := int32(0)
-	for _, output := range o {
-		total += output.FailedEntryCount
-	}
-	return total
-}
-
-func (svc *EventBridgeServiceImpl) DeployScheduleRules(ctx context.Context, rules ScheduleRules, optFns ...func(*eventbridge.Options)) (DeployScheduleRulesOutput, error) {
-	ret := make([]*DeployScheduleRuleOutput, 0, len(rules))
-	for _, rule := range rules {
-		output, err := svc.DeployScheduleRule(ctx, rule, optFns...)
-		if err != nil {
-			return nil, err
-		}
-		ret = append(ret, output)
-	}
-	return ret, nil
-}
-
-func (svc *EventBridgeServiceImpl) DeleteScheduleRule(ctx context.Context, rule *ScheduleRule, optFns ...func(*eventbridge.Options)) error {
-	targetIDs := make([]string, 0, len(rule.Targets))
-	for _, target := range rule.Targets {
-		targetIDs = append(targetIDs, *target.Id)
-	}
-	_, err := svc.client.RemoveTargets(ctx, &eventbridge.RemoveTargetsInput{
-		Ids:          targetIDs,
-		Rule:         rule.Name,
-		EventBusName: rule.EventBusName,
-	}, optFns...)
-	if err != nil {
-		return err
-	}
-	_, err = svc.client.DeleteRule(ctx, &eventbridge.DeleteRuleInput{
-		Name:         rule.Name,
-		EventBusName: rule.EventBusName,
-	}, optFns...)
-	return err
-}
-
-func (svc *EventBridgeServiceImpl) DeleteScheduleRules(ctx context.Context, rules ScheduleRules, optFns ...func(*eventbridge.Options)) error {
-	for _, rule := range rules {
-		if err := svc.DeleteScheduleRule(ctx, rule, optFns...); err != nil {
-			return fmt.Errorf("%s :%w", *rule.Name, err)
-		}
-	}
-	return nil
-}
-
-func (rule *ScheduleRule) SetStateMachineArn(stateMachineArn string) {
-	if rule.Description == nil {
-		rule.Description = aws.String(fmt.Sprintf("for state machine %s schedule", stateMachineArn))
-	}
-	if len(rule.Targets) == 0 {
-		rule.Targets = []eventbridgetypes.Target{
-			{
-				RoleArn: &rule.TargetRoleArn,
-			},
-		}
-	}
-	rule.Targets[0].Arn = aws.String(stateMachineArn)
-	if rule.Targets[0].Id == nil {
-		rule.Targets[0].Id = aws.String(fmt.Sprintf("%s-managed-state-machine", appName))
-	}
-}
-
-func (rule *ScheduleRule) IsManagedBy() bool {
-	for _, tag := range rule.Tags {
-		if *tag.Key == tagManagedBy && *tag.Value == appName {
-			return true
-		}
-	}
-	return false
-}
-
-func (rule *ScheduleRule) AppendTags(tags map[string]string) {
-	notExists := make([]eventbridgetypes.Tag, 0, len(tags))
-	aleradyExists := make(map[string]string, len(rule.Tags))
-	pos := make(map[string]int, len(rule.Tags))
-	for i, tag := range rule.Tags {
-		aleradyExists[*tag.Key] = *tag.Value
-		pos[*tag.Key] = i
-	}
-	for key, value := range tags {
-		if _, ok := aleradyExists[key]; !ok {
-			notExists = append(notExists, eventbridgetypes.Tag{
-				Key:   aws.String(key),
-				Value: aws.String(value),
-			})
-			continue
-		}
-		rule.Tags[pos[key]].Value = aws.String(value)
-	}
-	rule.Tags = append(rule.Tags, notExists...)
-}
-
-func (rule *ScheduleRule) configureJSON() string {
-	tags := make(map[string]string, len(rule.Tags))
-	for _, tag := range rule.Tags {
-		tags[*tag.Key] = *tag.Value
-	}
-	params := map[string]interface{}{
-		"Name":               rule.Name,
-		"Description":        rule.Description,
-		"ScheduleExpression": rule.ScheduleExpression,
-		"State":              rule.State,
-		"Targets":            rule.Targets,
-		"Tags":               tags,
-	}
-	return MarshalJSONString(params)
-}
-
-func (rule *ScheduleRule) String() string {
-	var builder strings.Builder
-	builder.WriteString(colorRestString(rule.configureJSON()))
-	return builder.String()
-}
-
-func (rule *ScheduleRule) DiffString(newRule *ScheduleRule) string {
-	var builder strings.Builder
-	builder.WriteString(colorRestString(JSONDiffString(rule.configureJSON(), newRule.configureJSON())))
-	return builder.String()
-}
-
-func (rule *ScheduleRule) SetEnabled(enabled bool) {
-	if enabled {
-		rule.State = eventbridgetypes.RuleStateEnabled
-	} else {
-		rule.State = eventbridgetypes.RuleStateDisabled
-	}
-}
-
-func (rules ScheduleRules) SetStateMachineArn(stateMachineArn string) {
-	for _, rule := range rules {
-		rule.SetStateMachineArn(stateMachineArn)
-	}
-}
-
-func (rules ScheduleRules) String() string {
-	var builder strings.Builder
-
-	for _, rule := range rules {
-		builder.WriteString(rule.String())
-		builder.WriteRune('\n')
-	}
-	return builder.String()
-}
-
-func (rules ScheduleRules) SetEnabled(enabled bool) {
-	for _, rule := range rules {
-		rule.SetEnabled(enabled)
-	}
-}
-
-func (rules ScheduleRules) SyncState(other ScheduleRules) {
-	otherMap := make(map[string]*ScheduleRule, len(other))
-
-	for _, r := range other {
-		name := ""
-		if r.Name != nil {
-			name = *r.Name
-		}
-		otherMap[name] = r
-	}
-	for _, r := range rules {
-		name := ""
-		if r.Name != nil {
-			name = *r.Name
-		}
-		if o, ok := otherMap[name]; ok {
-			r.State = o.State
-		}
-	}
-}
-
-// Things that are in rules but not in other
-func (rules ScheduleRules) Subtract(other ScheduleRules) ScheduleRules {
-	nothing := make(ScheduleRules, 0, len(rules))
-	otherMap := make(map[string]*ScheduleRule, len(other))
-	for _, r := range other {
-		otherMap[*r.Name] = r
-	}
-	for _, r := range rules {
-		if _, ok := otherMap[*r.Name]; !ok {
-			nothing = append(nothing, r)
-		}
-	}
-	return nothing
-}
-
-func (rules ScheduleRules) Exclude(other ScheduleRules) ScheduleRules {
-	otherMap := make(map[string]*ScheduleRule, len(other))
-	for _, r := range other {
-		otherMap[*r.Name] = r
-	}
-
-	ret := make(ScheduleRules, 0, len(rules))
-	ret = append(ret, rules...)
-	for i, r := range ret {
-		if _, ok := otherMap[*r.Name]; ok {
-			ret = append(ret[:i], ret[i+1:]...)
-		}
-	}
-	return ret
-}
-
-func (rules ScheduleRules) DiffString(newRules ScheduleRules) string {
-	addRuleName := make([]string, 0)
-	deleteRuleName := make([]string, 0)
-	changeRuleName := make([]string, 0)
-	ruleMap := make(map[string]*ScheduleRule, len(rules))
-	newRuleMap := make(map[string]*ScheduleRule, len(newRules))
-
-	for _, r := range newRules {
-		newRuleMap[*r.Name] = r
-	}
-	for _, r := range rules {
-		ruleMap[*r.Name] = r
-		if _, ok := newRuleMap[*r.Name]; ok {
-			changeRuleName = append(changeRuleName, *r.Name)
-		} else {
-			deleteRuleName = append(deleteRuleName, *r.Name)
-		}
-	}
-	for _, r := range newRules {
-		if _, ok := ruleMap[*r.Name]; !ok {
-			addRuleName = append(addRuleName, *r.Name)
-		}
-	}
-
-	var builder strings.Builder
-	for _, name := range deleteRuleName {
-		rule := ruleMap[name]
-		builder.WriteString(colorRestString(JSONDiffString(rule.configureJSON(), "null")))
-	}
-	for _, name := range changeRuleName {
-		rule := ruleMap[name]
-		newRule := newRuleMap[name]
-		builder.WriteString(rule.DiffString(newRule))
-	}
-	for _, name := range addRuleName {
-		newRule := newRuleMap[name]
-		builder.WriteString(colorRestString(JSONDiffString("null", newRule.configureJSON())))
-	}
-	return builder.String()
-}
